pyproject.toml
[project]
name = "ra"
version = "0.1.0"
description = "Robo Advisor Agent System with MCP and A2A"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.6.2",
    "fastmcp>=2.11.0",
    "a2a-sdk>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "faiss-cpu>=1.8.0",
    "streamlit>=1.39.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.1",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "langchain-mcp-adapters>=0.1.12",
    "tavily>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["a2a", "agents", "mcp_custom", "streamlit_app", "vector_db"]



[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
